requirements.txt
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
requests
pylint
autopep8
python-dateutil
inflect
pillow
docker